Update constants, move getPerturber state one level up
