+++ thor/orbits/ephemeris/__init__.py
@@ -1,4 +1,3 @@
 from .pyoorb import *
-from .spice import *
 from .universal import *
 from .ephemeris import *
