+++ thor/orbits/state.py
@@ -1,12 +1,12 @@
 import numpy as np
 import spiceypy as sp
 
-from ...constants import Constants as c
-from ...utils import setupSPICE
-from ...utils import _checkTime
+from ..constants import Constants as c
+from ..utils import setupSPICE
+from ..utils import _checkTime
 
-KM_TO_AU = c.KM_TO_AU
-S_TO_DAY = c.S_TO_DAY
+KM_P_AU = c.KM_P_AU
+S_P_DAY = c.S_P_DAY
 
 NAIF_MAPPING = {
     "solar system barycenter" : 0,
@@ -76,7 +76,7 @@ def getPerturberState(body_name, times, frame="ecliptic", origin="heliocenter"):
 
     # Convert MJD epochs in UTC to ET in TDB
     epochs_utc = times.utc
-    epochs_et = np.array([sp.str2et("JD {:.16f} UTC".format(i)) for i in epochs_utc.jd])
+    epochs_et = np.array([sp.str2et('JD {:.16f} UTC'.format(i)) for i in epochs_utc.jd])
     
     # Get position of the body in heliocentric ecliptic J2000 coordinates 
     states = []
@@ -92,6 +92,6 @@ def getPerturberState(body_name, times, frame="ecliptic", origin="heliocenter"):
     states = np.vstack(states)
     
     # Convert to AU and AU per day
-    states *= KM_TO_AU
-    states[:, 3:] = states[:, 3:] / S_TO_DAY
+    states = states / KM_P_AU
+    states[:, 3:] = states[:, 3:] * S_P_DAY
     return states
